Strip leading/trailing whitespace in input fields
gpacalc was broken for Firefox users because copying the "Grades at a
Glance" table on MySwat using Firefox resulted in columns separated by
[space][tab] while Chrome separates columns using only [tab].

This meant that each field contained an extra trailing space. Since
grades like "A " (with trailing space) do not match any of the grades in
the GRADE_POINT_EQUIVS dictionary, every course was ignored in GPA
calculation, resulting in the ZeroDivisionError that generates the
"Invalid grade input" error on submission.